engine/pyproject.toml
[tool.poetry]
name = "engine"
version = "0.1.0"
description = ""
authors = ["akita8 <mbianchi.akita@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
grpcio = "^1.31.0"
protobuf = "^3.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.8.3"
grpcio-tools = "^1.31.0"

[tool.poetry.scripts]
main = 'engine.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
